Перевел проект на kotlin, исправил имена пакетов с маленькой буквы, сконвертировал Note.java на kotlin
