--- app/src/main/java/project/paveltoy/noteapp/UI/NoteSourceCallback.java
@@ -1,7 +1,0 @@
-package project.paveltoy.noteapp.UI;
-
-import project.paveltoy.noteapp.Data.NoteRepository;
-
-public interface NoteSourceCallback {
-    void initialized(NoteRepository notes);
-}
