+++ app/src/main/java/project/paveltoy/noteapp/data/FirebaseAccountOpenData.java
@@ -1,4 +1,4 @@
-package project.paveltoy.noteapp.Data;
+package project.paveltoy.noteapp.data;
 
 import android.net.Uri;
