+++ app/src/main/java/project/paveltoy/noteapp/data/NoteMapping.java
@@ -1,4 +1,4 @@
-package project.paveltoy.noteapp.Data;
+package project.paveltoy.noteapp.data;
 
 import com.google.firebase.Timestamp;
