+++ app/src/main/java/project/paveltoy/noteapp/data/TemporaryClassNotes.java
@@ -1,4 +1,4 @@
-package project.paveltoy.noteapp.Data;
+package project.paveltoy.noteapp.data;
 
 import java.util.ArrayList;
 import java.util.List;
