+++ app/src/main/java/project/paveltoy/noteapp/ui/NoteAdapter.java
@@ -1,4 +1,4 @@
-package project.paveltoy.noteapp.UI;
+package project.paveltoy.noteapp.ui;
 
 import android.view.LayoutInflater;
 import android.view.View;
@@ -9,8 +9,8 @@ import android.widget.Filterable;
 import androidx.annotation.NonNull;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.RecyclerView;
-import project.paveltoy.noteapp.Data.Note;
-import project.paveltoy.noteapp.Data.NoteRepository;
+import project.paveltoy.noteapp.data.Note;
+import project.paveltoy.noteapp.data.NoteRepository;
 import project.paveltoy.noteapp.R;
 
 public class NoteAdapter extends RecyclerView.Adapter<NoteViewHolder> implements Filterable {
