+++ app/src/main/java/project/paveltoy/noteapp/ui/NoteListFragment.java
@@ -1,4 +1,4 @@
-package project.paveltoy.noteapp.UI;
+package project.paveltoy.noteapp.ui;
 
 import android.content.Context;
 import android.content.res.Configuration;
@@ -10,9 +10,9 @@ import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.GridLayoutManager;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
-import project.paveltoy.noteapp.Data.FirebaseNoteRepository;
-import project.paveltoy.noteapp.Data.Note;
-import project.paveltoy.noteapp.Data.NoteRepository;
+import project.paveltoy.noteapp.data.FirebaseNoteRepository;
+import project.paveltoy.noteapp.data.Note;
+import project.paveltoy.noteapp.data.NoteRepository;
 import project.paveltoy.noteapp.R;
 
 import android.view.ContextMenu;
