+++ app/src/main/java/project/paveltoy/noteapp/ui/NoteSourceCallback.java
@@ -1,0 +1,7 @@
+package project.paveltoy.noteapp.ui;
+
+import project.paveltoy.noteapp.data.NoteRepository;
+
+public interface NoteSourceCallback {
+    void initialized(NoteRepository notes);
+}
