+++ app/src/main/res/layout/fragment_list_notes.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".UI.NoteListFragment"
+    tools:context=".ui.NoteListFragment"
     android:orientation="vertical">
 
     <androidx.recyclerview.widget.RecyclerView
